+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.3] - 2018-10-03
 ### Added
 - Code of Conduct
 - CircleCI status badge
@@ -36,6 +38,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/WeInfuse/redox/compare/0.1.2...HEAD
+[Unreleased]: https://github.com/WeInfuse/redox/compare/0.1.3...HEAD
+[0.1.3]: https://github.com/WeInfuse/redox/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/WeInfuse/redox/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/WeInfuse/redox/compare/0.1.0...0.1.1